requirements.txt
# Core Airflow
apache-airflow==3.1.0
apache-airflow-core==3.1.0

# Airflow Providers
apache-airflow-providers-amazon==9.16.0
apache-airflow-providers-common-sql==1.28.2
apache-airflow-providers-postgres==6.4.0
apache-airflow-providers-snowflake==6.6.0
apache-airflow-providers-standard==1.9.1
apache-airflow-task-sdk==1.1.0

# Data Processing
pandas==2.3.3
pyarrow==22.0.0
numpy==2.3.4

# Database
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11
snowflake-connector-python==3.18.0

# AWS (for S3 operations via pyarrow)
boto3==1.40.58
botocore==1.40.58
pandas==2.3.3
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyarrow==22.0.0
python-dotenv==1.1.1
s3fs==2025.10.0
s3transfer==0.14.0
snowflake-connector-python==3.18.0
